CIF-2514: Split gift card into hook + component
