+++ react-components/src/index.js
@@ -21,6 +21,7 @@ export { default as AccountContainer } from './components/AccountContainer';
 export { default as AddressBook } from './components/AddressBook';
 export { default as BundleProductOptions } from './components/BundleProductOptions';
 export { default as GiftCartOptions } from './components/GiftCardOptions';
+export { default as useGiftCartOptions } from './components/GiftCardOptions/useGiftCardOptions';
 export { Portal } from './components/Portal';
 export { default as PortalPlacer } from './components/PortalPlacer';
 export { default as ResetPassword } from './components/ResetPassword';
